ai-dj/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
opencv-python==4.9.0.80
ultralytics>=8.3.0
deepface==0.0.93
tf-keras==2.16.0
spotipy==2.23.0
phue==1.1
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.3.0
